move hyper_cli_exec_test.go to passed
